refactor: introduce hexagonal architecture interfaces for AWS decoupling
Complete the clean separation of AWS dependencies from @friggframework/core
by introducing proper port/adapter interfaces following DDD principles.

New interfaces (ports) in core:
- QueueClientInterface: sendMessage, sendMessageBatch, getQueueUrl
- EncryptionKeyProviderInterface: generateDataKey, decryptDataKey
- WebSocketMessageSenderInterface: send (with StaleConnectionError)
- AesEncryptionKeyProvider: core adapter for local AES encryption

New adapters in @friggframework/provider-aws:
- SqsQueueClient: SQS implementation of QueueClientInterface
- KmsEncryptionKeyProvider: KMS implementation of EncryptionKeyProviderInterface
- ApiGatewayMessageSender: API Gateway impl of WebSocketMessageSenderInterface
- KMS health check: VPC detection + KMS capability check extracted from health.js

Updated core classes to use interfaces (all backward-compatible):
- Worker: accepts optional queueClient, lazy-loads SqsQueueClient as default
- Cryptor: accepts optional keyProvider, lazy-loads KMS or AES provider based
  on shouldUseAws flag (existing constructor API preserved)
- QueuerUtil: delegates to SqsQueueClient, adds setQueueClient() for DI
- WebSocket repos (4 Prisma variants + 1 Mongoose): accept optional
  messageSender, lazy-load ApiGatewayMessageSender as default
- health.js: lazy-loads KMS health check from provider-aws

Zero @aws-sdk imports remain in @friggframework/core source files.
All existing tests pass with no regressions.
