+++ packages/core/queues/queuer-util.js
@@ -1,52 +1,33 @@
 const { v4: uuid } = require('uuid');
 
-// AWS SQS SDK is lazy-loaded to avoid pulling in @aws-sdk/client-sqs
-// on non-AWS platforms. The singleton client is created on first use.
-let _sqsModule = null;
-let _sqs = null;
+// Queue client is lazy-loaded to avoid pulling in AWS SDK on non-AWS platforms.
+let _queueClient = null;
 
-function getSqsModule() {
-    if (!_sqsModule) {
-        _sqsModule = require('@aws-sdk/client-sqs');
+function getQueueClient() {
+    if (!_queueClient) {
+        const { SqsQueueClient } = require('@friggframework/provider-aws');
+        _queueClient = new SqsQueueClient();
     }
-    return _sqsModule;
-}
-
-const awsConfigOptions = () => {
-    const config = {};
-    if (process.env.IS_OFFLINE) {
-        config.credentials = {
-            accessKeyId: 'test-aws-key',
-            secretAccessKey: 'test-aws-secret',
-        };
-        config.region = 'us-east-1';
-    }
-    if (process.env.AWS_ENDPOINT) {
-        config.endpoint = process.env.AWS_ENDPOINT;
-    }
-    return config;
-};
-
-function getSqs() {
-    if (!_sqs) {
-        const { SQSClient } = getSqsModule();
-        _sqs = new SQSClient(awsConfigOptions());
-    }
-    return _sqs;
+    return _queueClient;
 }
 
 const QueuerUtil = {
+    /**
+     * Override the queue client (useful for testing or non-AWS platforms).
+     * @param {QueueClientInterface} client
+     */
+    setQueueClient(client) {
+        _queueClient = client;
+    },
+
     send: async (message, queueUrl) => {
-        const { SendMessageCommand } = getSqsModule();
-        const command = new SendMessageCommand({
+        return getQueueClient().sendMessage({
             MessageBody: JSON.stringify(message),
             QueueUrl: queueUrl,
         });
-        return getSqs().send(command);
     },
 
     batchSend: async (entries = [], queueUrl) => {
-        const { SendMessageBatchCommand } = getSqsModule();
         const buffer = [];
         const batchSize = 10;
 
@@ -57,11 +38,10 @@ const QueuerUtil = {
             });
             // Sends 10, then purges the buffer
             if (buffer.length === batchSize) {
-                const command = new SendMessageBatchCommand({
+                await getQueueClient().sendMessageBatch({
                     Entries: buffer,
                     QueueUrl: queueUrl,
                 });
-                await getSqs().send(command);
                 // Purge the buffer
                 buffer.splice(0, buffer.length);
             }
@@ -69,11 +49,10 @@ const QueuerUtil = {
 
         // If any remaining entries under 10 are left in the buffer, send and return
         if (buffer.length > 0) {
-            const command = new SendMessageBatchCommand({
+            return getQueueClient().sendMessageBatch({
                 Entries: buffer,
                 QueueUrl: queueUrl,
             });
-            return getSqs().send(command);
         }
 
         // If we're exact... just return an empty object for now
