+++ packages/core/types/core/index.d.ts
@@ -1,6 +1,4 @@
 declare module '@friggframework/core' {
-    import type { SendMessageCommandInput } from '@aws-sdk/client-sqs';
-
     export class Delegate implements IFriggDelegate {
         delegate: any;
         delegateTypes: any[];
@@ -26,7 +24,16 @@ declare module '@friggframework/core' {
         ): Promise<any>;
     }
 
+    /** Interface for queue message operations (port in hexagonal architecture) */
+    export interface QueueClientInterface {
+        sendMessage(params: SendMessageParams): Promise<{ MessageId: string }>;
+        sendMessageBatch(params: SendMessageBatchParams): Promise<any>;
+        getQueueUrl(params: GetQueueURLParams): Promise<string>;
+    }
+
     export class Worker implements IWorker {
+        constructor(options?: { queueClient?: QueueClientInterface });
+
         getQueueURL(params: GetQueueURLParams): Promise<string | undefined>;
 
         run(params: { Records: any }): Promise<void>;
@@ -36,7 +43,7 @@ declare module '@friggframework/core' {
             delay?: number
         ): Promise<string>;
 
-        sendAsyncSQSMessage(params: SendSQSMessageParams): Promise<string>;
+        sendAsyncSQSMessage(params: SendMessageParams): Promise<string>;
     }
 
     interface IWorker {
@@ -46,7 +53,7 @@ declare module '@friggframework/core' {
             params: object & { QueueUrl: any },
             delay?: number
         ): Promise<string>;
-        sendAsyncSQSMessage(params: SendSQSMessageParams): Promise<string>;
+        sendAsyncSQSMessage(params: SendMessageParams): Promise<string>;
     }
 
     export function loadInstalledModules(): any[];
@@ -56,5 +63,37 @@ declare module '@friggframework/core' {
         QueueOwnerAWSAccountId?: string;
     };
 
-    type SendSQSMessageParams = SendMessageCommandInput;
+    type SendMessageParams = {
+        QueueUrl: string;
+        MessageBody: string;
+        DelaySeconds?: number;
+    };
+
+    type SendMessageBatchParams = {
+        QueueUrl: string;
+        Entries: Array<{ Id: string; MessageBody: string }>;
+    };
+
+    /** Interface for envelope encryption key operations */
+    export class EncryptionKeyProviderInterface {
+        generateDataKey(): Promise<{
+            keyId: string;
+            encryptedKey: string;
+            plaintext: string | Buffer;
+        }>;
+        decryptDataKey(
+            keyId: string,
+            encryptedKey: string | Buffer
+        ): Promise<string | Buffer>;
+    }
+
+    /** Interface for WebSocket message sending */
+    export class WebSocketMessageSenderInterface {
+        send(connectionId: string, data: any, endpoint: string): Promise<void>;
+    }
+
+    export class StaleConnectionError extends Error {
+        connectionId: string;
+        constructor(connectionId: string);
+    }
 }
