+++ packages/providers/aws/index.js
@@ -28,6 +28,21 @@ module.exports = {
     get SqsQueueProvider() {
         return require('./queues/sqs-queue-provider').SqsQueueProvider;
     },
+    get SqsQueueClient() {
+        return require('./queues/sqs-queue-client').SqsQueueClient;
+    },
+
+    // ── Encryption ────────────────────────────────────────────────────
+    get KmsEncryptionKeyProvider() {
+        return require('./encryption/kms-encryption-key-provider')
+            .KmsEncryptionKeyProvider;
+    },
+
+    // ── WebSocket ─────────────────────────────────────────────────────
+    get ApiGatewayMessageSender() {
+        return require('./websocket/api-gateway-message-sender')
+            .ApiGatewayMessageSender;
+    },
 
     // ── Lambda ───────────────────────────────────────────────────────
     get LambdaInvoker() {
@@ -41,5 +56,13 @@ module.exports = {
     get MigrationStatusRepositoryS3() {
         return require('./storage/migration-status-repository-s3')
             .MigrationStatusRepositoryS3;
+    },
+
+    // ── Health ────────────────────────────────────────────────────────
+    get checkKmsDecryptCapability() {
+        return require('./health/kms-health-check').checkKmsDecryptCapability;
+    },
+    get detectVpcConfiguration() {
+        return require('./health/kms-health-check').detectVpcConfiguration;
     },
 };
